rename forwardEvent to trackEvent
