+++ packages/destination-actions/src/destinations/playerzero-cloud/trackEvent/__tests__/snapshot.test.ts
@@ -4,7 +4,7 @@ import destination from '../../index'
 import nock from 'nock'
 
 const testDestination = createTestIntegration(destination)
-const actionSlug = 'forwardEvent'
+const actionSlug = 'trackEvent'
 const destinationSlug = 'PlayerzeroCloud'
 const seedName = `${destinationSlug}#${actionSlug}`
