+++ packages/destination-actions/src/destinations/playerzero-cloud/util.ts
@@ -1,6 +1,6 @@
 import { RequestClient } from '@segment/actions-core/dist/cjs'
 import { Settings } from './generated-types'
-import { Payload } from './forwardEvent/generated-types'
+import { Payload } from './trackEvent/generated-types'
 
 export const send = async (request: RequestClient, settings: Settings, payload: Payload[], type: string) => {
   return request(`https://sdk.playerzero.app/connect/segment/${type}`, {
